Move createTestApp to it's own file
