+++ test/demo/createNotification.inttest.ts
@@ -4,11 +4,11 @@ import {Application, Identity} from 'hadouken-js-adapter';
 
 import {Notification, NotificationOptions} from '../../src/client';
 
-import {fin} from './utils/fin';
 import * as notifsRemote from './utils/notificationsRemoteExecution';
 import {getCardsByNotification, isCenterShowing} from './utils/notificationCenterUtils';
 import {delay} from './utils/delay';
 import {getToastWindow, getToastCards} from './utils/toastUtils';
+import {createTestApp} from './utils/createTestApp';
 
 const validOptions: NotificationOptions = {
     body: 'Test Notification Body',
@@ -143,24 +143,3 @@ describe('When calling createNotification', () => {
         });
     });
 });
-
-// TODO: Make this a util modeled on layouts' spawn utils (SERVICE-524)
-const nextUuid = (() => {
-    let count = 0;
-    return () => 'notifications-test-app-' + (count++);
-})();
-
-async function createTestApp():Promise<Application> {
-    const uuid = nextUuid();
-    const app = await fin.Application.create({
-        uuid,
-        name: uuid,
-        url: 'http://localhost:3922/test/test-app.html',
-        autoShow: true,
-        showTaskbarIcon: false,
-        defaultHeight: 400,
-        defaultWidth: 500
-    });
-    await app.run();
-    return app;
-}
