+++ test/demo/eventListeners.inttest.ts
@@ -3,10 +3,10 @@ import {Application, Window as FinWindow} from 'hadouken-js-adapter';
 
 import {NotificationClickedEvent, Notification, NotificationOptions, NotificationButtonClickedEvent, NotificationClosedEvent} from '../../src/client';
 
-import {fin} from './utils/fin';
 import * as notifsRemote from './utils/notificationsRemoteExecution';
 import {getCardsByNotification, isCenterShowing} from './utils/notificationCenterUtils';
 import {delay} from './utils/delay';
+import {createTestApp} from './utils/createTestApp';
 
 const defaultNoteOptions: NotificationOptions = {
     body: 'Test Notification Body',
@@ -146,25 +146,3 @@ describe('Click listeners', () => {
         });
     });
 });
-
-
-// TODO: Make this a util modeled on layouts' spawn utils (SERVICE-524)
-const nextUuid = (() => {
-    let count = 100;
-    return () => 'notifications-test-app-' + (count++);
-})();
-
-async function createTestApp():Promise<Application> {
-    const uuid = nextUuid();
-    const app = await fin.Application.create({
-        uuid,
-        name: uuid,
-        url: 'http://localhost:3922/test/test-app.html',
-        autoShow: true,
-        showTaskbarIcon: false,
-        defaultHeight: 400,
-        defaultWidth: 500
-    });
-    await app.run();
-    return app;
-}
